# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wardrobe AI API",
    "description": "REST API for Wardrobe AI. Import this file into Postman via **Import → File → openapi.json**.",
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:3000/api/v1", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "example": "user@example.com" },
          "mobile": { "type": "string", "example": "+919876543210" },
          "password": { "type": "string", "minLength": 8, "example": "SecurePass123!" }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "email": { "type": "string", "example": "user@example.com" },
          "mobile": { "type": "string", "example": "+919876543210" },
          "password": { "type": "string", "example": "SecurePass123!" }
        }
      },
      "RefreshTokenDto": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string" }
        }
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "gender": { "type": "string", "enum": ["MALE", "FEMALE", "OTHER", "PREFER_NOT_TO_SAY"] },
          "age": { "type": "integer", "example": 28 },
          "height": { "type": "number", "example": 165 },
          "weight": { "type": "number", "example": 62 },
          "body_type": { "type": "string" },
          "skin_tone": { "type": "string" }
        }
      },
      "CreateProductDto": {
        "type": "object",
        "required": ["sku", "name", "category_id", "brand_id", "price"],
        "properties": {
          "sku": { "type": "string", "example": "SKU-001" },
          "name": { "type": "string", "example": "Classic Denim Jacket" },
          "description": { "type": "string" },
          "category_id": { "type": "string" },
          "brand_id": { "type": "string" },
          "price": { "type": "number", "example": 79.99 },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": { "type": "string" },
                "sort_order": { "type": "integer" },
                "is_primary": { "type": "boolean" }
              }
            }
          }
        }
      },
      "AddToWishlistDto": {
        "type": "object",
        "required": ["product_id"],
        "properties": {
          "product_id": { "type": "string", "example": "product-uuid" }
        }
      },
      "CreateOrderDto": {
        "type": "object",
        "required": ["total_amount"],
        "properties": {
          "total_amount": { "type": "number", "example": 149.99 }
        }
      },
      "FaceEmbeddingDto": {
        "type": "object",
        "required": ["embedding"],
        "properties": {
          "embedding": {
            "type": "array",
            "items": { "type": "number" },
            "example": [0.12, -0.05, 0.88]
          }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["health"],
        "summary": "Health check",
        "responses": { "200": { "description": "Service is healthy" } }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterDto" } }
          }
        },
        "responses": { "201": { "description": "User registered" }, "409": { "description": "Email or mobile exists" } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login with email or mobile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginDto" } }
          }
        },
        "responses": { "200": { "description": "Login successful" }, "401": { "description": "Invalid credentials" } }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Refresh access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenDto" } }
          }
        },
        "responses": { "200": { "description": "Tokens refreshed" } }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Logout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenDto" } }
          }
        },
        "responses": { "200": { "description": "Logout successful" } }
      }
    },
    "/users/profile": {
      "get": {
        "tags": ["users"],
        "summary": "Get authenticated user profile",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Profile retrieved" }, "401": { "description": "Unauthorized" } }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update authenticated user profile",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProfileDto" } }
          }
        },
        "responses": { "200": { "description": "Profile updated" } }
      }
    },
    "/products": {
      "get": {
        "tags": ["products"],
        "summary": "List products",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "category_id", "in": "query", "schema": { "type": "string" } },
          { "name": "brand_id", "in": "query", "schema": { "type": "string" } },
          { "name": "min_price", "in": "query", "schema": { "type": "number" } },
          { "name": "max_price", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Products list" } }
      },
      "post": {
        "tags": ["products"],
        "summary": "Create product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateProductDto" } }
          }
        },
        "responses": { "201": { "description": "Product created" } }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["products"],
        "summary": "Get product by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Product found" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["products"],
        "summary": "Update product",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProductDto" } } }
        },
        "responses": { "200": { "description": "Product updated" } }
      },
      "delete": {
        "tags": ["products"],
        "summary": "Delete product",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Product deleted" } }
      }
    },
    "/recommendations": {
      "get": {
        "tags": ["recommendations"],
        "summary": "Get personalized recommendations",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }],
        "responses": { "200": { "description": "Recommendations list" } }
      }
    },
    "/fashion-dna": {
      "get": {
        "tags": ["fashion-dna"],
        "summary": "Get Fashion DNA",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Fashion DNA retrieved" } }
      }
    },
    "/fashion-dna/generate": {
      "post": {
        "tags": ["fashion-dna"],
        "summary": "Generate Fashion DNA",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Fashion DNA generated" } }
      }
    },
    "/wishlist": {
      "get": {
        "tags": ["wishlist"],
        "summary": "Get wishlist",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Wishlist retrieved" } }
      },
      "post": {
        "tags": ["wishlist"],
        "summary": "Add product to wishlist",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AddToWishlistDto" } }
          }
        },
        "responses": { "201": { "description": "Added to wishlist" } }
      }
    },
    "/wishlist/{id}": {
      "delete": {
        "tags": ["wishlist"],
        "summary": "Remove wishlist item",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Item removed" } }
      }
    },
    "/orders": {
      "get": {
        "tags": ["orders"],
        "summary": "List user orders",
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Orders list" } }
      },
      "post": {
        "tags": ["orders"],
        "summary": "Create order",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrderDto" } }
          }
        },
        "responses": { "201": { "description": "Order created" } }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["orders"],
        "summary": "Get order by ID",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Order found" } }
      }
    },
    "/face/register": {
      "post": {
        "tags": ["face"],
        "summary": "Register face embedding",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/FaceEmbeddingDto" } }
          }
        },
        "responses": { "201": { "description": "Face registered" } }
      }
    },
    "/face/login": {
      "post": {
        "tags": ["face"],
        "summary": "Login with face embedding",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/FaceEmbeddingDto" } }
          }
        },
        "responses": { "200": { "description": "Face login successful" } }
      }
    }
  },
  "tags": [
    { "name": "health" },
    { "name": "auth" },
    { "name": "users" },
    { "name": "products" },
    { "name": "recommendations" },
    { "name": "fashion-dna" },
    { "name": "wishlist" },
    { "name": "orders" },
    { "name": "face" }
  ]
}